Change links to unsupported versions
Change links to unsupported versions. This relative link is consistent with other links in our other docs. Starting a link with `/docs/` and adding the `.md` extension forced the page to be shown as a raw Markdown file, so I had removed the extension to work around that issue at the time. This change fixes that.